LSystemRuleを外部ファイルへ移動
