+++ src/simulations/drawer/model.ts
@@ -2,7 +2,8 @@ import p5 from "p5"
 import { Vector } from "../../classes/physics"
 import { random } from "../../classes/utilities"
 import { Drawer } from "./drawer"
-import { LSystemDrawer, LSystemRule } from "./lsystem_drawer"
+import { LSystemRule } from "./lsystem_rule"
+import { LSystemDrawer } from "./lsystem_drawer"
 import { Line, isCollided } from "./line"
 import { QuadtreeNode } from "./quadtree"
 // Do not import constants (pass constants via Model.constructor)
